ServerlessDayFunctions2019/requirements.txt
tensorflow==1.12.0
pillow==5.0.0
numpy==1.14.1